refactor(cookbook): split detail-data into per-recipe files with flexible structure
- Extract RecipeDetail and related types to data/types.ts
- Split 18 recipes into data/recipes/*.ts (one file per recipe)
- Create data/index.ts as the new barrel (RECIPE_DETAILS + getRecipeDetail)
- Make principle, rules, pattern, implementation optional in RecipeDetail
- Add rulesLabel field to customize the section heading per recipe
- Update CookbookDetailPage to render sections conditionally with
  dynamic section numbering (SectionBadge component)
- Update all 6 import locations to use @/features/cookbook/data
- Delete old detail-data.ts (1939 lines → 18 focused files)
